fix: form validation fixes, avatars update, move logic from hooks to controller
