+++ src/containers/Profiles/avatarUrls.json
@@ -1,12 +1,14 @@
 [
-  "https://gravatar.com/avatar/5e62c8c13582f94b74ae21cfeb83e28a?s=400&d=robohash&r=x",
-  "https://gravatar.com/avatar/a82dc2482b1ae8d9070462a37b5e19e9?s=400&d=robohash&r=x",
-  "https://gravatar.com/avatar/236030198309afe28c9fce9c3ebfec3b?s=400&d=robohash&r=x",
-  "https://gravatar.com/avatar/c97a042d43cc5cc28802f2bc7bf2e5ab?s=400&d=robohash&r=x",
-  "https://gravatar.com/avatar/5e62c8c13582f94b74ae21c34b83e28a?s=400&d=robohash&r=x",
-  "https://gravatar.com/avatar/a82dc2482b1ae8d9070462a3723e19e9?s=400&d=robohash&r=x",
-  "https://gravatar.com/avatar/5e62c8c13582f94b74ae21c34b43e28a?s=400&d=robohash&r=x",
-  "https://gravatar.com/avatar/5e62c8c13582f94b74ae21c24b83e28a?s=400&d=robohash&r=x",
-  "https://gravatar.com/avatar/5e62c8c13582f94b74ae21c14b83e28a?s=400&d=robohash&r=x",
-  "https://gravatar.com/avatar/5e62c8c13582f94b74aw21c34b83e28a?s=400&d=robohash&r=x"
+  "/images/avatars/Alien.svg",
+  "/images/avatars/Bear.svg",
+  "/images/avatars/Brainy.svg",
+  "/images/avatars/Cooool.svg",
+  "/images/avatars/Dummy.svg",
+  "/images/avatars/Frog.svg",
+  "/images/avatars/Goofball.svg",
+  "/images/avatars/Marilyn.svg",
+  "/images/avatars/Smiley.svg",
+  "/images/avatars/ToughGuy.svg",
+  "/images/avatars/UhOh.svg",
+  "/images/avatars/Vibe.svg"
 ]
