+++ src/stores/ProfileController.ts
@@ -1,5 +1,17 @@
-import useService from '#src/hooks/useService';
+import { useFavoritesStore } from './FavoritesStore';
+import { useProfileStore } from './ProfileStore';
+import { useWatchHistoryStore } from './WatchHistoryStore';
+import { initializeAccount } from './AccountController';
+
+import * as persist from '#src/utils/persist';
 import type { ProfilePayload, EnterProfilePayload, ProfileDetailsPayload } from '#types/account';
+import useService from '#src/hooks/useService';
+
+const PERSIST_PROFILE = 'profile';
+
+export const unpersistProfile = () => {
+  persist.removeItem(PERSIST_PROFILE);
+};
 
 export const listProfiles = async () => {
   return await useService(async ({ profileService, sandbox }) => {
@@ -21,7 +33,20 @@ export const updateProfile = async ({ id, name, adult, avatar_url, pin }: Profil
 
 export const enterProfile = async ({ id, pin }: EnterProfilePayload) => {
   return await useService(async ({ profileService, sandbox }) => {
-    return await profileService?.enterProfile({ id, pin }, sandbox ?? true);
+    const response = await profileService?.enterProfile({ id, pin }, sandbox ?? true);
+    const profile = response?.responseData;
+    if (profile?.credentials?.access_token) {
+      persist.setItem(PERSIST_PROFILE, profile);
+      persist.setItemStorage('inplayer_token', {
+        expires: profile.credentials.expires,
+        token: profile.credentials.access_token,
+        refreshToken: '',
+      });
+      useFavoritesStore.setState({ favorites: [] });
+      useWatchHistoryStore.setState({ watchHistory: [] });
+      useProfileStore.getState().setProfile(profile);
+    }
+    return initializeAccount();
   });
 };
